credential: rename stackplan/stackcred package
