+++ go/src/koding/kites/kloud/credential/socialapi.go
@@ -1,4 +1,4 @@
-package stackcred
+package credential
 
 import (
 	"bytes"
